requirements.txt
django==4.1.7
djangorestframework
psycopg2-binary